General Dicrete Cellular Automata

CellularAutomata.DCA — Type

DCA(rule; states=2, radius=1)

Returns a DCA object given a specific rule, number of states and radius. The ruleset for the rule is computed and stored in the struct as well.

Totalistic Discrete Cellular Automata

CellularAutomata.TCA — Type

(tca::TCA)(starting_array)

Returns the next state of the given starting_array according to the evolution rule contained in the TCA struct.

Continous Cellular Automata

CellularAutomata.CCA — Type

(cca::CCA)(starting_array)

Returns the next state of the given starting_array according to the evolution rule contained in the CCA struct.