One dimensional Cellular Automata

Discrete Cellular Automata

CellularAutomata.DCA — Type

DCA(rule; states=2, radius=1)

Returns a DCA object given a specific rule, number of states and radius. The ruleset for the rule is computed and stored in the struct as well.

CellularAutomata.TCA — Type

TCA(code; states=2, radius=1)

Returns a TCA object given a specific code, number of states and radius. The ruleset for the rule is computed and stored in the struct as well.

Continuous Cellular Automata

CellularAutomata.CCA — Type

TCA(code; radius=1)

Returns a CCA object given a specific code and radius.